services/customer-service/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
motor==3.3.1
pymongo==4.6.1
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
dnspython==2.4.2
boto3==1.34.34
email-validator==2.1.0.post1 